requirements.txt
altgraph==0.17.4
certifi==2025.4.26
charset-normalizer==3.4.2
docopt==0.6.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pipreqs==0.4.13
PyGetWindow==0.0.9
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyRect==0.2.0
pywin32==310
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==80.3.1
urllib3==2.4.0
yarg==0.1.10
